summary nell'admin

+++ backend/src/main/java/com/fantasta/dto/BidDto.java
@@ -2,5 +2,5 @@ package com.fantasta.dto;
 
 public class BidDto {
     public Long participantId;
-    public int amount;
+    public Double amount;
 }
